==> 01.Preprocessing.ipynb <==
import os
import json
import re
import numpy as np
import pandas as pd
import sqlite3 as sql
import seaborn as sns
import difflib

### 1. Product

products = pd.read_csv("Cung cap HV/ProductRaw.csv")
products.head(2)

products.info()

# Number of products
print('The number of products:', products.shape[0])

# Drop duplicate records
print('Trước khi dropping, records =', products.shape[0])
products.drop_duplicates(inplace=True)
print('Sau khi dropping, records =', products.shape[0])

# Check null value
products.isnull().any()

# Drop null value
# print('Trước khi dropping, records =', products.shape[0])
# products.dropna(inplace=True)
# print('Sau khi dropping, records =', products.shape[0])

# Save result
products.to_csv('Cung cap HV/Product_new.csv', index=False)

### 2. Review

reviews = pd.read_csv("Cung cap HV/ReviewRaw.csv")
reviews.head()

# Data type
reviews.info()

# Number of reviews
print('Number of reviews:', reviews.shape[0])

# Check null value
reviews.isnull().any()

m = []
for c in ['name','full_name','created_time']:
    m.append({
        'feature':c,
        'MissingVal': reviews[reviews[c].isnull()].shape[0],
        'Percentage': (reviews[reviews[c].isnull()].shape[0]/reviews.shape[0])*100
    })
df = pd.DataFrame(m)
df

Nhận xét:
* created_time có dữ liệu null khá lớn
* full_name có dữ liệu null xấp xỉ 9.5%

# Drop feature created_time and full_name
reviews.drop(['full_name','created_time'],axis=1, inplace=True)

reviews.shape

# Save result
reviews.to_csv('Cung cap HV/Review_new.csv', index=False)

### 3. Product - Review

products = pd.read_csv("Cung cap HV/Product_new.csv")
reviews = pd.read_csv("Cung cap HV/Review_new.csv")

# Show products in reviews dataset that is not in products dataset
reviews[~reviews.product_id.isin(products.item_id)]

# Drop reviews that have non-existent product_id
reviews = reviews[reviews.product_id.isin(products.item_id)]

reviews.reset_index(drop=True,inplace=True)

reviews.shape[0]

reviews.head()

# Save result
reviews.to_csv('Cung cap HV/Review_new.csv',index=False)
products.to_csv('Cung cap HV/Product_new.csv', index=False)
