==> 3.Content_Based_Filtering_Gensim.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import linear_kernel, cosine_similarity
from underthesea import word_tokenize, pos_tag, sent_tokenize
import warnings
from gensim import corpora, models, similarities
import jieba
import re

warnings.filterwarnings('ignore')

### Load data

products = pd.read_csv("Cung cap HV/Product_new.csv")
reviews = pd.read_csv("Cung cap HV/Review_new.csv")

products.head(2)

products.info()

products.shape

### Data Transformation

products = products[products['name'].notnull()]

products['name_description'] = products.name + products.description

products = products[products['name_description'].notnull()]

products['name_desciption_pre'] = products['name_description'].apply(lambda x: word_tokenize(x,format='text'))

type(products)

products.shape

### Solution 2: Gensim

intro_products = [[text for text in x.split()] for x in products.name_desciption_pre]

len(intro_products)

intro_products[:1]

# remove ' ' element in texts
intro_products_re = [[t.lower() for t in text if not t in ['', ' ', ',', '.', '...', '-',':', ';', '?', '%', '(', ')', '+', '\d+','...']] for text in  intro_products]

intro_products_re[:1]

# Obtain the number of features based on dictionary: Use corpora.Dictionary
dictionary = corpora.Dictionary(intro_products_re)

# List of features in dictionary
dictionary.token2id

# Number of features (words) in dictionaries
feature_count = len(dictionary.token2id)

feature_count

# Obtain corpus based on dictionary (dense matrix)
corpus = [dictionary.doc2bow(text) for text in intro_products_re]

corpus[0]

# Use TF-IDF Model to process corpus, obtaining index
tfidf = models.TfidfModel(corpus)
# tính toán sự tương tự trong ma trận thưa thớt
index = similarities.SparseMatrixSimilarity(tfidf[corpus],num_features=feature_count)

# When user choose one product: 1059892, 52333193
product_ID = 10001355
product = products[products.item_id == product_ID].head(1)

type(product['name_desciption_pre'])

product[['item_id','name_desciption_pre']]

# View product
name_description_pre = product['name_desciption_pre'].to_string(index=False)

name_description_pre

# Suggest other products for customers
def recommendation (view_product, dictionary, tfidf, index):
    # Convert search words into Sparse Vectors
    view_product = view_product.lower().split()
    kw_vector = dictionary.doc2bow(view_product)
    print("View product 's vector:")
    print(kw_vector)
    # Similarity calculation
    sim = index[tfidf[kw_vector]]
    
    # print result
    list_id = []
    list_score = []
    for i in range(len(sim)):
        list_id.append(i)
        list_score.append(sim[i])
    
    df_result = pd.DataFrame({'id': list_id,
                              'score': list_score})
    
    # five highest scores
    five_highest_score = df_result.sort_values(by='score', ascending=False).head(6)
    print("Five highest scores:")
    print(five_highest_score)
    print("Ids to list:")
    idToList = list(five_highest_score['id'])
    print(idToList)
    
    products_find = products[products.index.isin(idToList)]
    results = products_find[['item_id','name']]
    results = pd.concat([results, five_highest_score], axis=1).sort_values(by='score', ascending=False)
    return results

results = recommendation(name_description_pre, dictionary, tfidf, index)

# Recommender 5 similarities products for the selected product
# Check and remove the selected product from the results
results = results[results.item_id!=product_ID]
results

### Save results

dictionary.save("Gensim_Model.sav")

dictionary.load("Gensim_Model.sav")
