==> 4.Collaborative_Filtering_ALS.ipynb <==
import findspark
findspark.init()

from pyspark.sql import SparkSession
from pyspark import SparkContext
from pyspark.sql import SQLContext
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.ml.recommendation import ALS
from pyspark.sql.functions import col, explode

SparkContext.setSystemProperty('spark.executor.memory','12g')
sc = SparkContext(master='local', appName='Recommendation')

spark = SparkSession(sc)

# data1 = spark.read.csv('Cung cap HV/Product_new.csv',header=True,inferSchema=True)
data = spark.read.csv('Cung cap HV/Review_new.csv',header=True,inferSchema=True)

data.show(5, truncate=True)

data_sub = data.select(['product_id','rating','customer_id'])

data.count()

data_sub.printSchema()

from pyspark.sql.types import DoubleType, IntegerType
from pyspark.sql.functions import isnan, when, count, col, udf

data_sub = data_sub.withColumn('rating', data_sub["rating"].cast(DoubleType()))
data_sub = data_sub.withColumn('customer_id', data_sub["customer_id"].cast(IntegerType()))
data_sub = data_sub.withColumn('product_id', data_sub["product_id"].cast(IntegerType()))

data_sub.select([count(when(col(c).isNull(), c)).alias(c) for c in 
           data_sub.columns]).toPandas().T

data_sub = data_sub.na.drop(how='any')

data_sub.count()

data_sub.select([count(when(col(c).isNull(), c)).alias(c) for c in 
           data_sub.columns]).toPandas().T

# Distinct users and products
users = data_sub.select("customer_id").distinct().count()
products = data_sub.select("product_id").distinct().count()
numerator = data_sub.count()

display(numerator, users, products)

# Number of rating matrix could contain if no empty cells
denominator = users * products
denominator

sparsity = 1- (numerator * 1.0/ denominator)
print("sparsity:"), sparsity

### Feature Transformation

from pyspark.ml.feature import StringIndexer
from pyspark.ml import Pipeline
from pyspark.sql.functions import col

indexer = StringIndexer(inputCol='product_id',
                        outputCol='product_id_idx')
data_indexed = indexer.fit(data_sub).transform(data_sub)

indexer1 = StringIndexer(inputCol='customer_id',
                        outputCol='customer_id_idx')
data_indexed = indexer1.fit(data_indexed).transform(data_indexed)

data_indexed.show(5, truncate=True)

### Train model

# Chia dữ liệu train_test
# Smaller dataset so we will use 0.8/0.2
(training, test)= data_indexed.randomSplit([0.8,0.2])

# xây dựng model
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.ml.recommendation import ALS

als = ALS(maxIter=10,
          regParam=0.09,
          rank=25,
          userCol="customer_id_idx",
          itemCol="product_id_idx",
          ratingCol="rating",
          coldStartStrategy="drop",
          nonnegative=True)
model = als.fit(training)
# rank cho to hơn vì phim nhiều và user nhiều

data_indexed.select("rating").describe().show()

### Evaluate model

predictions = model.transform(test)

predictions.show(5)

from pyspark.ml.evaluation import RegressionEvaluator

evaluator = RegressionEvaluator(metricName='rmse',
                                labelCol="rating",
                                predictionCol="prediction")
rmse = evaluator.evaluate(predictions)
print("Root-mean square error = "+ str(rmse))

### Tunning parameter

ALSExplicit = ALS( implicitPrefs=False, userCol="customer_id_idx", itemCol="product_id_idx", ratingCol="rating",
          coldStartStrategy="drop")

defaultModel = ALSExplicit.fit(training)

from pyspark.ml.tuning import CrossValidator, ParamGridBuilder, TrainValidationSplit

paramMapExplicit = ParamGridBuilder() \
                    .addGrid(ALSExplicit.rank, [30, 40,50]) \
                    .addGrid(ALSExplicit.maxIter, [5,10,15,20]) \
                    .addGrid(ALSExplicit.regParam, [0.1,0.01,0.001]) \
                    .addGrid(ALSExplicit.alpha, [2.0,3.0]) \
                    .build()

evaluatorR = RegressionEvaluator(metricName="rmse", labelCol="rating")

CVALSExplicit = CrossValidator(estimator=ALSExplicit,
                            estimatorParamMaps=paramMapExplicit,
                            evaluator=evaluatorR,
                           numFolds=5)


CVModelEXplicit = CVALSExplicit.fit(training)

predictions_t = CVModelEXplicit.transform(test)

rmse_t = evaluator.evaluate(predictions_t)
print("Root-mean square error = "+ str(rmse_t))

als_t = ALS(maxIter=10,
          regParam=0.1,
          rank = 30,
          userCol="customer_id_idx",
          itemCol="product_id_idx",
          ratingCol="rating",
          coldStartStrategy="drop",
          nonnegative=True)
model_t = als_t.fit(training)

predictions_t = model_t.transform(test)

rmse_t = evaluator.evaluate(predictions_t)
print("Root-mean square error = "+ str(rmse_t))

Nhận xét: Tốt hơn vì rmse nhỏ hơn nên sẽ sử dụng model này

### Make recommendations to all users

# Get 5 recommendations which have highest rating
user_recs = model_t.recommendForAllUsers(5)

user_recs.show(10,truncate=False)

user_recs.printSchema()

for user in user_recs.head(3):
    print(user)
    print("\n")

### Save to file

from time import time

t0 = time()

df_reviewer_reviewer_id = data_indexed.select('customer_id_idx', 'customer_id').distinct()

df_reviewer_reviewer_id.count()

df_reviewer_reviewer_id.show(5)

df_asin_asin_id = data_indexed.select('product_id_idx', 'product_id').distinct()

df_asin_asin_id.count()

df_asin_asin_id.show(5)

new_user_recs = user_recs.join(df_reviewer_reviewer_id, on=['customer_id_idx'], how="left")

new_user_recs.show(10, truncate=False)

new_user_recs.count()

### Save to disk

new_user_recs.write.parquet('Recommendation_U.parquet', mode='overwrite')
df_asin_asin_id.write.parquet('Recommendation_P.parquet', mode='overwrite')

time_duration = time() - 10
print(time_duration)

### Make recommendations to a particular user

customer_id = "5917275"
find_user_rec = new_user_recs.filter(new_user_recs['customer_id'] == customer_id)
user = find_user_rec.first()

lst=[]
for row in user['recommendations']:
    row_f = df_asin_asin_id.filter(df_asin_asin_id.product_id_idx == row['product_id_idx'])
    row_f_first = row_f.first()
    lst.append((row['product_id_idx'], row_f_first['product_id'], row['rating']))
dic_user_rec = {'customer_id' : user.customer_id, 'recommendations':lst}

dic_user_rec

### Đọc 2 file đã lưu để lấy dữ liệu đầu vào => Đề xuất

new_user_recs = spark.read.parquet('Recommendation_U.parquet')

new_user_recs.printSchema()

new_user_recs.show(2)

df_asin_asin_id = spark.read.parquet('Recommendation_P.parquet')

df_asin_asin_id.show(2)

customer_id = "5917275"
find_user_rec = new_user_recs.filter(new_user_recs['customer_id'] == customer_id)
find_user_rec.show(truncate=False)

result = ''
for user in find_user_rec.collect():
    lst=[]
    for row in user['recommendations']:
        print(row)
        row_f = df_asin_asin_id.filter(df_asin_asin_id.product_id_idx == row['product_id_idx'])
        row_f_first = row_f.first()
        lst.append((row['product_id_idx'], row_f_first['product_id'], row['rating']))
    dic_user_rec = {'customer_id' : user.customer_id, 'recommendations':lst}
    result = dic_user_rec
