==> 4.Collaborative_Filtering_Surprise.ipynb <==
import pandas as pd
import numpy as np
from surprise import Reader, Dataset, SVD, SVDpp, NMF, SlopeOne, KNNBasic, KNNBaseline, KNNWithMeans, KNNWithZScore, CoClustering, BaselineOnly
from surprise.model_selection.validation import cross_validate
from surprise import accuracy

data = pd.read_csv('Cung cap HV/Review_new.csv')
data.head()

data.info()

n_ratings = len(data)
n_products = len(data['product_id'].unique())
n_customers = len(data['customer_id'].unique())

display(n_ratings, n_products, n_customers)

data['product_id'].value_counts()

reader = Reader()
data_new = Dataset.load_from_df(data[['customer_id','product_id','rating']], reader)

from statistics import mean
def cal_avg_metrics(result):    
    result['test_rmse'] = result['test_rmse'].mean()
    result['test_mae'] = result['test_mae'].mean()    
    result['fit_time'] = round(mean(list(result['fit_time'])),3)
    result['test_time'] = round(mean(list(result['test_time'])),3)
    return result

#SVD, SVDpp, NMF, SlopeOne, KNNBasic, KNNBaseline, KNNWithMeans, KNNWithZScore, CoClustering, BaselineOnly

algorithms=[SVD()]
results=[]
names=[]

for algo in algorithms:
    result = cross_validate(algo, data_new, measures = ['RMSE', 'MAE'], cv=5, verbose=True) 
    result = cal_avg_metrics(result)
    results.append(result)
    names.append(algo.__class__.__name__)
df=pd.DataFrame(results)
df.insert(0, 'Algorithm', pd.Series(names))
df.head()

### Tunning Parameter: chỉ dành cho SVD

param_grid = {'n_epochs': [5, 10], 'lr_all': [0.002, 0.005], 'reg_all': [0.4, 0.6]}

from surprise.model_selection import GridSearchCV
gs = GridSearchCV(SVD, param_grid, measures=['rmse', 'mae'], cv=3)
gs.fit(data_new)

# best RMSE score
print(gs.best_score['rmse'])

# combination of parameters that gave the best RMSE score
print(gs.best_params['rmse'])

### Train model

model = BaselineOnly()

trainset = data_new.build_full_trainset()
model.fit(trainset)

### Recommender for a specific user

def recommendation (customerid, algorithm, df):
    df_score = df[['product_id']]
    df_score['EstimateScore'] = df_score['product_id'].apply(lambda x: algorithm.predict(customerid, x).est) 
    df_score = df_score.sort_values(by=['EstimateScore'], ascending=False)
    df_score = df_score.drop_duplicates()
    df_score = df_score[df_score['EstimateScore']>=3][:5]
    return df_score

customer_id = 5917275
df_new = recommendation(customer_id, model, data)

df_new
